pyproject.toml
[project]
name = "langywrap"
version = "0.1.0"
description = "Universal AI agent orchestration toolkit — execution security, ralph loops, hyperagents, compound engineering"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pyyaml",
    "click",
    "pydantic",
]

[project.optional-dependencies]
lean = []
data = [
    "duckdb",
    "polars",
]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "pytest-xdist",
]

[project.scripts]
langywrap = "langywrap.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib/langywrap"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
pretty = false

[tool.pytest.ini_options]
addopts = "-q --tb=short"
